refactor: move utils and fonts to new modules folder
