+++ jsconfig.json
@@ -4,6 +4,7 @@
     "paths": {
       "@/components/*": ["src/components/*"],
       "@/lib/*": ["src/lib/*"],
+      "@/modules/*": ["src/modules/*"],
       "@/providers/*": ["src/providers/*"],
       "@/styles/*": ["src/styles/*"]
     }
