+++ src/components/common/accordion.js
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import * as AccordionPrimitive from "@radix-ui/react-accordion";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 import { ChevronDown as ChevronDownIcon } from "@/components/common/icons";
 
 const Accordion = AccordionPrimitive.Root;
