+++ src/components/common/dropdown-menu.js
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 import {
   Check as CheckIcon,
   ChevronRight as ChevronRightIcon,
