+++ src/components/common/navigation-menu.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu";
 import { cva } from "class-variance-authority";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 import { ChevronDown as ChevronDownIcon } from "@/components/common/icons";
 
 const NavigationMenu = React.forwardRef(
