+++ src/components/common/separator.js
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import * as SeparatorPrimitive from "@radix-ui/react-separator";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 
 const Separator = React.forwardRef(
   (
