+++ src/components/common/sheet.js
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as SheetPrimitive from "@radix-ui/react-dialog";
 import { cva } from "class-variance-authority";
 import { X } from "lucide-react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 
 const Sheet = SheetPrimitive.Root;
