+++ src/components/common/typography.js
@@ -3,7 +3,7 @@
 import * as React from "react";
 import { Slot } from "@radix-ui/react-slot";
 import { cva } from "class-variance-authority";
-import { cn } from "@/lib/utils";
+import { cn } from "@/modules/utils";
 
 const typographyComponents = {
   h1: "h1",
